fix: LSP fallback behavior when formatters not availble
The bug would present itself when there were formatters configured for a
filetype that were not available (not installed, or some other issue).
This would also happen if the filetype configured an empty list of
formatters.

The fix was to make sure that we fall back to LSP formatting _either_ if
the filetype had no formatters configured, _or_ if none of the
configured formatters are available.