fix(prettier): do not run prettier on any Node project
Removed the cwd root file "package.json" for the prettier formatter. Prettier should be used to format code only when there is prettierrc-type file present, and not also when there is a package.json file present. This allows projects that use other formatters such as dprint or eslint with various formatter integration plugins to to use their preferred formatter without prettier overwriting their output as can be easily seen with projects created with create-effect-app for example, which prefers dprint.