fix: change LSP fallback logic for wildcard filetypes
Previous behavior: if you have "_" or "*" formatters configured AND you
use `lsp_format = "fallback"`, those formatters would be ignored in
favor of formatting with the LSP.

New behavior: there is no special ignore logic. Now the "_"/"*"
formatters will run instead of the LSP.

Why?
You can now set a custom `lsp_format` option inside the "_" wildcard
itself. For example:

```lua
conform.formatters_by_ft = {
  ["_"] = { "trim_whitespace", lsp_format = "last" }
}
```

This is a more elegant and readable solution for the previous problem.
What was the previous problem? There was no way to specify "for matching
filetypes, run specific formatters when available and fall back to LSP,
BUT for non-matching filetypes _always_ run some formatters _and_ the
LSP if it is available."