fix(csharpier): use csharpier 1.0.0 format command (#695)
* fix(csharpier)!: use csharpier format command

Since the 1.0.0 release of csharpier, the CLI has changed and the format
command is required.

The new arguments for the CLI require csharpier
>= 1.0.0.

* fix(csharpier): allow csharpier globally and locally

The csharpier invocation varies depending on whether the tool is
installed locally or globally. We favor local installation, and if it
isn't present we fallback on the global installation.