fix(treefmt): require config file
Running `treefmt` when no config file exists just results in an error,
so checking for the presence of a config file by default makes sense.

This will allow `treefmt` to be included as a formatter that should be
run on all file types, without throwing errors for projects that have no
treefmt config file.